+++ Magic/CHANGELOG.md
@@ -15,6 +15,7 @@
  - Added "damage" and "damage_dealt" attributes
  - Add "scope" to ModifyVariable, can be cast (the default), spell or mage
  - Attributes can now be used in Message and Command actions
+ - Add "direction" and "use_target" parameters to CheckBlock action, for easily checking the block under your feet
 
 # 7.9.1